refactor(slm-backend): extract _extract_failure_summary to ansible_utils (#5564)
Move the Ansible stdout parser from setup_wizard.py to a shared
services/ansible_utils.py module and apply it in infrastructure.py
and tls.py so all three callers surface human-readable failure
summaries instead of raw exit codes.
