+++ autobot-slm-backend/api/setup_wizard.py
@@ -11,7 +11,6 @@ node addition, enrollment, role assignment, and fleet provisioning.
 
 import asyncio
 import logging
-import re
 import tempfile
 import time
 import uuid
@@ -27,6 +26,7 @@ from pydantic import BaseModel
 
 from api.websocket import ws_manager
 from services.ansible_secrets import fetch_deploy_secrets
+from services.ansible_utils import _extract_failure_summary
 from services.auth import get_current_user
 from services.database import db_service
 from services.playbook_executor import get_playbook_executor
@@ -642,56 +642,6 @@ def _write_provision_log(line: str) -> None:
             f.write(line + "\n")
     except OSError:
         pass
-
-
-def _extract_failure_summary(output: str) -> str:
-    """Parse Ansible stdout and return a human-readable failure summary.
-
-    Extracts failed hosts, the task that failed, and the error message so
-    users see e.g. '172.16.168.26 failed at "Common | Update apt cache":
-    Failed to update apt cache: unknown reason' instead of 'exit code 2'.
-    """
-    lines = output.splitlines()
-    failures: list[str] = []
-    current_task = ""
-
-    i = 0
-    while i < len(lines):
-        line = lines[i].strip()
-
-        # Track the current task name
-        if line.startswith("TASK ["):
-            task_match = re.search(r"TASK \[(.+?)\]", line)
-            if task_match:
-                current_task = task_match.group(1).strip()
-
-        # Fatal failure or unreachable
-        if line.startswith("fatal:"):
-
-            host_match = re.search(r"fatal: \[([^\]]+)\]", line)
-            host = host_match.group(1) if host_match else "unknown host"
-            failure_type = "UNREACHABLE" if "UNREACHABLE" in line else "FAILED"
-
-            # Look ahead for msg: field in the next few lines
-            msg = ""
-            for j in range(i + 1, min(i + 10, len(lines))):
-                msg_match = re.search(r'"?msg"?\s*[:=]\s*["\']?(.+?)["\']?\s*$', lines[j].strip())
-                if msg_match:
-                    msg = msg_match.group(1).strip().strip("'\"")
-                    break
-
-            task_part = f' at "{current_task}"' if current_task else ""
-            msg_part = f": {msg}" if msg else ""
-            failures.append(f"{host} {failure_type.lower()}{task_part}{msg_part}")
-
-        i += 1
-
-    if not failures:
-        return ""
-
-    count = len(failures)
-    noun = "host" if count == 1 else "hosts"
-    return f"{count} {noun} failed — " + "; ".join(failures)
 
 
 def _handle_provision_result(result: dict) -> None:
